Now the program makes an attempt to read WINDOWS-1252 encoded text from zips
