+++ Cargo.toml
@@ -13,8 +13,7 @@ zip = "0.5"
 clap = {version = "2", optional = true}
 serde = {version = "1.0", features = ["derive"]}
 regex = "1"
-encoding_rs = "0.8"
-encoding_rs_io = "0.1"
+encoding = "0.2"
 diff_match_patch = "0.1"
